chore: h2 db용 Entity 폴더 경로 변경
