+++ src/main/java/study/toy/everythingshop/controller/SampleController.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import study.toy.everythingshop.entity.SampleEntity;
+import study.toy.everythingshop.entity.h2.SampleEntity;
 import study.toy.everythingshop.service.SampleService;
 
 import java.util.List;
