+++ src/main/java/study/toy/everythingshop/repository/ProductDAO.java
@@ -3,8 +3,7 @@ package study.toy.everythingshop.repository;
 import study.toy.everythingshop.dto.ProductOrderDTO;
 import study.toy.everythingshop.dto.ProductRegisterDTO;
 import study.toy.everythingshop.dto.ProductSearchDTO;
-import study.toy.everythingshop.entity.ProductMEntity;
-import study.toy.everythingshop.logTrace.Trace;
+import study.toy.everythingshop.entity.h2.ProductMEntity;
 
 import java.util.List;
