+++ src/main/java/study/toy/everythingshop/repository/impl/SampleDAOImpl.java
@@ -3,7 +3,7 @@ package study.toy.everythingshop.repository.impl;
 import lombok.RequiredArgsConstructor;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.stereotype.Repository;
-import study.toy.everythingshop.entity.SampleEntity;
+import study.toy.everythingshop.entity.h2.SampleEntity;
 import study.toy.everythingshop.repository.SampleDAO;
 
 import java.util.List;
