+++ src/main/java/study/toy/everythingshop/service/MyPageService.java
@@ -2,7 +2,7 @@ package study.toy.everythingshop.service;
 
 import study.toy.everythingshop.dto.ProductOrderDTO;
 import study.toy.everythingshop.dto.ProductSearchDTO;
-import study.toy.everythingshop.entity.UserMEntity;
+import study.toy.everythingshop.entity.h2.UserMEntity;
 
 import java.util.List;
