+++ src/main/java/study/toy/everythingshop/service/SampleService.java
@@ -1,6 +1,6 @@
 package study.toy.everythingshop.service;
 
-import study.toy.everythingshop.entity.SampleEntity;
+import study.toy.everythingshop.entity.h2.SampleEntity;
 
 import java.util.List;
