+++ src/main/java/study/toy/everythingshop/service/UserService.java
@@ -1,7 +1,7 @@
 package study.toy.everythingshop.service;
 
 import study.toy.everythingshop.dto.JoinDTO;
-import study.toy.everythingshop.entity.UserMEntity;
+import study.toy.everythingshop.entity.h2.UserMEntity;
 
 import java.util.List;
 public interface UserService {
