+++ src/test/java/study/toy/everythingshop/sample/SampleDAOTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import study.toy.everythingshop.entity.SampleEntity;
+import study.toy.everythingshop.entity.h2.SampleEntity;
 import study.toy.everythingshop.repository.SampleDAO;
 
 import java.util.List;
